// generated_new/question/model/question-set.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface QuestionSet {
    'id'?: string;
    'title'?: string;
    'description'?: string;
    'tags'?: string;
    'questionNum'?: number;
    'userId'?: string;
    'favourNum'?: number;
    'createTime'?: string;
    'updateTime'?: string;
    'isDelete'?: number;
}