// generated_new/question/model/question-set-query-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface QuestionSetQueryRequest {
    'current'?: number;
    'pageSize'?: number;
    'sortField'?: string;
    'sortOrder'?: string;
    'id'?: string;
    'title'?: string;
    'description'?: string;
    'userId'?: string;
}